resources/labs/python_sinewave/requirements.txt
pillow~=8.0
